# Parcial3/Swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API LEC 2023",
    "version": "1.0.0",
    "description": "API para el League Of Legends EMEA Chanpionship 2023"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/lec2023": {
      "get": {
        "summary": "Obtiene todos los equipos",
        "responses": {
          "200": {
            "description": "Retorna la lista de equipos"
          },
          "500": {
            "description": "Error en el servidor"
          }
        }
      },
      "post": {
        "summary": "Crea un nuevo equipo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string"
                  },
                  "acronimo": {
                    "type": "string"
                  },
                  "pais": {
                    "type": "string"
                  }
                },
                "required": ["nombre", "acronimo", "pais"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Equipo creado correctamente"
          },
          "400": {
            "description": "Faltan campos obligatorios en la solicitud"
          }
        }
      }
    },
    "/lec2023/{idEquipo}": {
      "get": {
        "summary": "Obtiene un equipo por su ID",
        "parameters": [
          {
            "in": "path",
            "name": "idEquipo",
            "required": true,
            "description": "ID del equipo a obtener",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna el equipo solicitado"
          },
          "404": {
            "description": "El equipo no fue encontrado"
          }
        }
      },
      "delete": {
        "summary": "Elimina un equipo por su ID",
        "parameters": [
          {
            "in": "path",
            "name": "idEquipo",
            "required": true,
            "description": "ID del equipo a eliminar",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Equipo eliminado correctamente"
          },
          "404": {
            "description": "El equipo no fue encontrado"
          }
        }
      },
      "put": {
        "summary": "Actualiza un equipo por su ID",
        "parameters": [
          {
            "in": "path",
            "name": "idEquipo",
            "required": true,
            "description": "ID del equipo a actualizar",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string"
                  },
                  "acronimo": {
                    "type": "string"
                  },
                  "pais": {
                    "type": "string"
                  }
                },
                "required": ["nombre", "acronimo", "pais"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Equipo actualizado correctamente"
          },
          "400": {
            "description": "Faltan campos obligatorios en la solicitud"
          },
          "404": {
            "description": "El equipo no fue encontrado"
          }
        }
      }
    }
  }
}